requirements.txt
streamlit>=1.36.0
lyricsgenius>=3.0.1
google-genai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
